improve eta method and add get distance
also change the type of distance in SQL from decimal to double

+++ sql-script.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS schedule(
 	ID INT AUTO_INCREMENT PRIMARY KEY,
     destination VARCHAR(20) NOT NULL,
     departure DATETIME NOT NULL,
-    distance DECIMAL(5, 2) NOT NULL,
+    distance DOUBLE NOT NULL,
     price DECIMAL(5, 2) NOT NULL # 3 digits before the decimal, 2 after
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=latin1;
